routes: extract static_routes (#187 third slice) (v5.95.3)
Third slice. Extracts the static-routes resource — handlers + the
crate-internal helpers the OPNsense importer and the backup-restore
path depend on.

What landed:

- `routes/static_routes.rs` (315 lines) — `StaticRoute`,
  `CreateRouteRequest`, `SystemRoute` + handlers (`list`, `create`,
  `update`, `delete`, `apply_all_routes`) + helpers (`validate_route_target`,
  `apply_route_to_system`, `remove_route_from_system`).
- mod.rs re-exports `validate_route_target` and `apply_route_to_system`
  with their original `pub(crate)` visibility so the `crate::routes::*`
  paths in `backup.rs`, `main.rs`, and `opnsense/importer.rs` keep
  working untouched.

`get_system_routes` (the netstat-backed read-only view) stays in mod.rs
for now since it's only ~40 lines and not on the static-route CRUD path.

Progress on #187 split so far:

| | mod.rs lines |
|---|---|
| Original | 4013 |
| After slice 1 (#266: dns + geoip + schedules) | 3689 |
| After slice 2 (#267: rules + nat) | 3160 |
| **After this PR (static_routes)** | **2858 (-1155 / -29%)** |

Six focused modules now own their resource. The same pattern fits the
remaining: auth, users+roles, vpn (WG + IPsec), interfaces, settings
group, backup/import_config, status/connections/logs/metrics. Each
can land in a follow-up commit.

Verified:
- `cargo check --workspace` clean
- `cargo test --package aifw-api` — 123 tests pass
